libs/py-normalize/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "synheart-normalize"
version = "0.1.0"
description = "Data normalization library for Synheart wearable integrations"
authors = [
    {name = "Israel Goytom", email = "israel@synheart.ai"}
]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "pydantic>=2.0.0",
    "python-dateutil>=2.8.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.3.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["synheart_normalize*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.ruff]
line-length = 100
target-version = "py311"
